// src/api/products/api.ts
export interface CreateProductRequest {
    /**
     * 
     * @type {string}
     * @memberof CreateProductRequest
     */
    'name': string;
    /**
     * 
     * @type {number}
     * @memberof CreateProductRequest
     */
    'price': number;
    /**
     * 
     * @type {string}
     * @memberof CreateProductRequest
     */
    'observation'?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateProductRequest
     */
    'expiresAt'?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof CreateProductRequest
     */
    'taxes'?: Array<string>;
    /**
     * 
     * @type {Array<string>}
     * @memberof CreateProductRequest
     */
    'discounts'?: Array<string>;
}
/**
 * 
 * @export
 * @interface ErrorResponse
 */
export interface ErrorResponse {
    /**
     * 
     * @type {string}
     * @memberof ErrorResponse
     */
    'error'?: string;
    /**
     * 
     * @type {number}
     * @memberof ErrorResponse
     */
    'code'?: number;
}
/**
 * 
 * @export
 * @interface GetAllProductsPaginatedResponse
 */
export interface GetAllProductsPaginatedResponse {
    /**
     * 
     * @type {Pagination}
     * @memberof GetAllProductsPaginatedResponse
     */
    'pagination'?: Pagination;
    /**
     * 
     * @type {Array<ProductResponse>}
     * @memberof GetAllProductsPaginatedResponse
     */
    'data'?: Array<ProductResponse>;
}
/**
 * 
 * @export
 * @interface Pagination
 */
export interface Pagination {
    /**
     * 
     * @type {number}
     * @memberof Pagination
     */
    'totalItems'?: number;
    /**
     * 
     * @type {number}
     * @memberof Pagination
     */
    'pageSize'?: number;
    /**
     * 
     * @type {number}
     * @memberof Pagination
     */
    'currentPage'?: number;
    /**
     * 
     * @type {number}
     * @memberof Pagination
     */
    'totalPages'?: number;
}
/**
 * 
 * @export
 * @interface ProductResponse
 */
export interface ProductResponse {
    /**
     * 
     * @type {string}
     * @memberof ProductResponse
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof ProductResponse
     */
    'reference'?: string;
    /**
     * 
     * @type {string}
     * @memberof ProductResponse
     */
    'name'?: string;
    /**
     * 
     * @type {number}
     * @memberof ProductResponse
     */
    'price'?: number;
    /**
     * 
     * @type {string}
     * @memberof ProductResponse
     */
    'observation'?: string;
    /**
     * 
     * @type {string}
     * @memberof ProductResponse
     */
    'expiresAt'?: string;
    /**
     * 
     * @type {string}
     * @memberof ProductResponse
     */
    'createdAt'?: string;
    /**
     * 
     * @type {number}
     * @memberof ProductResponse
     */
    'taxes'?: number;
    /**
     * 
     * @type {number}
     * @memberof ProductResponse
     */
    'descount'?: number;
}
/**
 * 
 * @export
 * @interface UpdateProductRequest
 */
export interface UpdateProductRequest {
    /**
     * 
     * @type {string}
     * @memberof UpdateProductRequest
     */
    'name'?: string;
    /**
     * 
     * @type {number}
     * @memberof UpdateProductRequest
     */
    'price'?: number;
    /**
     * 
     * @type {string}
     * @memberof UpdateProductRequest
     */
    'observation'?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateProductRequest
     */
    'expiresAt'?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof UpdateProductRequest
     */
    'taxes'?: Array<string>;
    /**
     * 
     * @type {Array<string>}
     * @memberof UpdateProductRequest
     */
    'discounts'?: Array<string>;
}
// ...
export interface DefaultApiCreateProductRequest {
    /**
     * 
     * @type {CreateProductRequest}
     * @memberof DefaultApiCreateProduct
     */
    readonly createProductRequest: CreateProductRequest
}

/**
 * Request parameters for deleteProduct operation in DefaultApi.
 * @export
 * @interface DefaultApiDeleteProductRequest
 */
export interface DefaultApiDeleteProductRequest {
    /**
     * Unique identifier of the product to delete
     * @type {string}
     * @memberof DefaultApiDeleteProduct
     */
    readonly id: string
}

/**
 * Request parameters for getAllProducts operation in DefaultApi.
 * @export
 * @interface DefaultApiGetAllProductsRequest
 */
export interface DefaultApiGetAllProductsRequest {
    /**
     * The starting index of the products.
     * @type {number}
     * @memberof DefaultApiGetAllProducts
     */
    readonly from?: number

    /**
     * The number of products to return.
     * @type {number}
     * @memberof DefaultApiGetAllProducts
     */
    readonly size?: number

    /**
     * Filter criteria for products.
     * @type {string}
     * @memberof DefaultApiGetAllProducts
     */
    readonly filter?: string

    /**
     * Field to order products by.
     * @type {'createdAt'}
     * @memberof DefaultApiGetAllProducts
     */
    readonly orderBy?: GetAllProductsOrderByEnum

    /**
     * Sorting direction.
     * @type {'ASC' | 'DESC'}
     * @memberof DefaultApiGetAllProducts
     */
    readonly direction?: GetAllProductsDirectionEnum

    /**
     * Filter product by their status.
     * @type {'ACTIVE' | 'INACTIVE'}
     * @memberof DefaultApiGetAllProducts
     */
    readonly status?: GetAllProductsStatusEnum
}

/**
 * Request parameters for updateProductPartially operation in DefaultApi.
 * @export
 * @interface DefaultApiUpdateProductPartiallyRequest
 */
export interface DefaultApiUpdateProductPartiallyRequest {
    /**
     * The ID of the product to update
     * @type {string}
     * @memberof DefaultApiUpdateProductPartially
     */
    readonly id: string

    /**
     * 
     * @type {UpdateProductRequest}
     * @memberof DefaultApiUpdateProductPartially
     */
    readonly updateProductRequest: UpdateProductRequest
}

/**
 * DefaultApi - object-oriented interface
 * @export
 * @class DefaultApi
 * @extends {BaseAPI}
 */
export class DefaultApi extends BaseAPI {
    /**
     * 
     * @summary Create a new product
     * @param {DefaultApiCreateProductRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DefaultApi
     */
    public createProduct(requestParameters: DefaultApiCreateProductRequest, options?: RawAxiosRequestConfig) {
        return DefaultApiFp(this.configuration).createProduct(requestParameters.createProductRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Delete a product by ID
     * @param {DefaultApiDeleteProductRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DefaultApi
     */
    public deleteProduct(requestParameters: DefaultApiDeleteProductRequest, options?: RawAxiosRequestConfig) {
        return DefaultApiFp(this.configuration).deleteProduct(requestParameters.id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Retrieve a list of products
     * @param {DefaultApiGetAllProductsRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DefaultApi
     */
    public getAllProducts(requestParameters: DefaultApiGetAllProductsRequest = {}, options?: RawAxiosRequestConfig) {
        return DefaultApiFp(this.configuration).getAllProducts(requestParameters.from, requestParameters.size, requestParameters.filter, requestParameters.orderBy, requestParameters.direction, requestParameters.status, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Partially update a product
     * @param {DefaultApiUpdateProductPartiallyRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DefaultApi
     */
    public updateProductPartially(requestParameters: DefaultApiUpdateProductPartiallyRequest, options?: RawAxiosRequestConfig) {
        return DefaultApiFp(this.configuration).updateProductPartially(requestParameters.id, requestParameters.updateProductRequest, options).then((request) => request(this.axios, this.basePath));
    }
}

/**
 * @export
 */
export const GetAllProductsOrderByEnum = {
    CreatedAt: 'createdAt'
} as const;
export type GetAllProductsOrderByEnum = typeof GetAllProductsOrderByEnum[keyof typeof GetAllProductsOrderByEnum];
/**
 * @export
 */
export const GetAllProductsDirectionEnum = {
    Asc: 'ASC',
    Desc: 'DESC'
} as const;
export type GetAllProductsDirectionEnum = typeof GetAllProductsDirectionEnum[keyof typeof GetAllProductsDirectionEnum];
/**
 * @export
 */
export const GetAllProductsStatusEnum = {
    Active: 'ACTIVE',
    Inactive: 'INACTIVE'
} as const;
export type GetAllProductsStatusEnum = typeof GetAllProductsStatusEnum[keyof typeof GetAllProductsStatusEnum];